mcp-server/requirements.txt
mcp>=1.0.0
openai>=1.30.0
supabase>=2.4.0
python-dotenv>=1.0.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
tiktoken>=0.7.0
starlette>=0.40.0
